refactor(decimal_serde): simplify module structure by removing unnecessary module declaration
